Still use nbin when process.send exists
In code-server we need to restart the process when upgrading but we also
need IPC to communicate with the child and this made it so we can't do
that without specifying a path to the main file which is not stable
because the path can be different between binaries (at the very least
the version will be different but there's no guarantee the rest of the
path is the same either).

This change makes it so you can control whether nbin is used for any
spawn including with IPC, but it also means you need to be sure you set
NBIN_BYPASS appropriately (code-server sets it immediately upon
spawning to ensure nothing spawns/forks with nbin unless we explicitly
pass `NBIN_BYPASS: undefined` to it).