Get nbin working again
It now only loads the third party main file when NBIN_BYPASS isn't set
and that file then directly loads the main file. This is mostly because
I couldn't quite figure out how to get it working the way it used to due
to structural Node changes but I think this might be more idiomatic
anyway.